refactor(offline): Extract useMinuteTick for the relative-time refresh
Changes:
- Add hooks/use-minute-tick.ts, encapsulating the mount gate and the 60s re-render
- Use it in LastSyncedLabel instead of an inline mounted flag and a discarded tick state

The sweep had shortened the comment covering `const [, setTick] = useState(0)`
plus its interval, leaving a discarded state value and a timer that read as dead
code. Naming the behaviour in a hook removes the need to explain it at the call
site, and the SSR-mismatch rationale now lives on the hook it belongs to.

Notes:
- Review finding #37
